Completed adding in multiple dataloaders...
to correct some N+1 query issues. tested and running endpoints.

small glitch to figure out. the authorsliceLoader_gen has a hiccup not
sure why. it imports as just Author everywhere when it needs to be
pg.Author.... replaceing Author with pg.Author everywhere fixes it.
will look into this later. probabbly an issue with the go generate line.